chore: separate index.ts from memory-storage.ts
